Prevent the command tab from closing and make it unique
At first glance, this seems like an unnecessary cut into the flexibility
of the frame functionality and this might be reverted some time in the
future.

But for now, it is easier to have a single dedicated command tab, so it
is easier to associate locals to the environment in the tab (models!).
Yes, the same reasoning goes for global variables, and they are not a
good idea. I will overthink this, if I ever see a concrete reason.