Support mixin initial conditions
Chosing for mixins not only because this is easier to implement than
persistent Segment.before_twiss hooks - but also because the user should
be in control when the DVM plugin reads values from the database and it
just doesn't make a lot of sense to dedicate an external handler which
doesn't even access external values. Also, keeping this responsibility
within MadGUI makes it much easier to maintain.

Resolves #70