Remove model.Locator class
When models were part of cpymad and multiple underlying storage engines
were supported the Locator class was needed to provide an extensible
mechanism to iterate and load models.

Now models have migrated to MadGUI and the only supported storage
backend is the local filesystem. Therefore this mechanism has become
unnecessary and can be removed for simplicity.