Add units configuration from current pint upstream
In the pint 0.6 release some units had their symbol and name exchanged,
notably 'eV' was showing up as 'electron_volt' in the dialog.

Note that the use of pint in madgui is not compatible with the current
pint upstream master, which is why I decided to just include the file
for now rather than just referring to the new upcoming version.