Fix mismatch between FloorPlan/TwissFigure selections
FloorPlan was using `expanded_elements` while selection via TwissFigure
was based on the unexpanded `elements` list. This causes a cascade of
exceptions when selecting an auto-generated DRIFT in the floorplan.

Therefore, we now always use `expanded_elements`. This has the advantage
that it shows exactly one element for each data point in TWISS/SURVEY
and allows to get rid of the extra `engine.survey_elements()` method.