Use IDs rather than names for selections
NOTE: IDs are currently just the element indices, but this should be
considered an implementation detail.

This change is necessary since element names in `madx.expanded_elements`
are not unique, i.e. there can be several auto-generated DRIFTs with
exactly the same MAD-X internal name.

An alternative possible solution is to reinstate uniqueness of element
names at an intermediate level (within cpymad), but this is awkward with
the current cpymad API: it would require to either:

    - recompute the external name of each auto-generated DRIFTs by
      counting preceding DRIFTs with the same name on every access

    - keep a separate cache of element names on the `cpymad.libmadx`
      or the `cpymad.madx.Madx` level – which conflicts with the current
      stateless API therefore requires big changes (this might be a good
      idea anyway!)

Also make sure to use a cpymad commit later than:

    Commit:  7762e2370681806e69d93e24afb18e5d3da4ff6f
    Date:    Mon Nov 14 19:49:31 2016 +0100
    Subject: Provide element index within the sequence