Rework data model for param dialogs
We replace pre-emtive data acquisition by so called `DataStore`s which
are responsible for mitigating recurring data retrieval and updates.

Although not reflected in the GUI, this modification also endorses
incremental (immediate/non-transactional) value updates - which means
that you now have a chance of setting values for the MAD-X elements.

The whole thing is still inherently broken and overcomplicated, but this
is a big step in a better direction (I'm not going so far as to say the
"right" direction…)

Big issues with this are still

- most parameters can't be written without crashing the program
- read/write decisions broken, especially for MAD-X
- GUI still acts as it were transactional (Ok/Apply/Cancel buttons)
- have to know the section, in order to do updates [tao]
- doesn't use the units provided by tao