Reduce the number of TWISS calls for sectormaps
By moving responsibility for caching the sectormaps to the controller
class we can get all transfer maps in one shot.

Also this makes the caching more dynamic by the following argument:
Since the MonitorItem are namedtuple, they are read-only. This means
that whenever we want to update the cache, we had to recreate the
MonitorItem and hence re-match - which can quickly become expensive.

With the new approach, we can invalidate the cache without having to
rematch.