Fix "jump-to-top" issue, more sensible data responsibilities
The SuperStore mechanism as used previously resulted in recreating all
the tabs and tableviews whenever the data changed or another element was
selected, thus creating flickering and resetting the view at the top.
After this patch, all widgets are kept once created.

Furthermore, we no longer generate all the tabs in the model, but rather
have the UI code directly in charge over what it wants to show.

Note there is still simplifications to be done. In particular, I dislike
the indirection introduced by the current usage of the DataStore model.