Replace all Expression instances by their values
This fixes a bug when showing expressions in TableView: The widget would
display only the expression, not its value -> leading to an inconsistent
look. Furthemore, when trying to edit the value in the TableView, there
was an error due to QuantitySpinBox being broken for Expression.

This problem was introduced in

    8775a999c06e69e8789e48bb66d1db60cf8ebfe6
    "Keep values as plain floats internally (in MAD-X units)"

i.e. when we started to use un-qualified values rather than unit-aware
quantities internally. Before this, all expressions were converted to
SymbolicValue which worked correctly with TableView.