Quick-fix hack to make multi_grid work…
By overwriting the element attribute expression each time before getting
the list of knobs, and a second time before matching.

Long-term this is partially correct: we need to overwrite the expression
before matching, but let the list of knobs be defined directly by the
model configuration.

The reason that we need to overwrite the expression each time is simple:
in order to get the correct model for the current sectormaps, we have to
get the exact kicker strengths from the online control. However, this is
in general impossible to do without removing the preexisting expression,
considering the kicks may not be of the form dictated by the expression.