Allow attribute access and changed signals on config sections
This generalizes the NumberFormat class which will later be removed in a
later commit.

Note that we only wrap config sections (dict) for now, but leave config
entries (scalar) as is for now. This has no particular reason other than
that I don't see the need yet to go all the way.

If we should wish to wrap all entries in the future, we should overload
the __call__ operator to allow getting/setting self._value.