Always reset model to force index invalidation
Without the patch, passing the TableCell (or its id) to `createIndex()`
as `internalPointer`, the view is unusable for edits after the second
update. This is likely due to the fact that the raw pointer value of
`internalPointer` counts toward equality comparison if QModelIndex. If
the view keeps indices with the old pointer value and mixes them with
indices with the new pointer this can result in failing to identify the
correct items.