Require max-complexity=11
I want to eventually bring this down to at most 10. Currently, the only
method with a complexity 11 is `OffsetCalibrationWidget.poll`, but it
should be relatively easy to simplify below 10 by replacing the
`prepare` flag by `shot == 0` criterium. Furthermore, we should simplify
and deduplicate this code a lot further with `ProcBot` anyway.