Lose obsolete thread utils [core.worker, QueuedDispatcher]
While I learnt a bit during the creation of these tools and grew quite
fond of them, it is now time to let go as they have become mostly
obsolete. The background thread for fetching data was only really
necessary for the pytao backend (where fetching data from the
interpreter would take enormous amounts of time due to the text
protocol).

With the cpymad backend, the situation is much better and we don't need
background threads for now.