Fix BeamOptikDLL not initiating device download (too flooded?)
It appears that BeamOptikDLL won't initiate device download after
execution if our process continuously calls GetFloatValueSD, or maybe it
is simply due to the process busy state.

The 10ms seems to be a typo introduced in a refactoring commit earlier,
it was 300ms before, but we don't even need to be that fast.