Change class Curve to plot_curve function
This brings a notable reduction of complexity, is a step forward in the
unified handling of `twiss_curves`/`user_curves` and also unifies the
`constr_markers` logic with other scene nodes.

This also removes the `update` feature that was meant to improve
performance when redrawing twiss curves and may also prevent other
unwanted behaviour (e.g. autoscaling). However, it seems that these
effects are currently hardly if at all noticeable.

In fact, this hints that we may fully remove `update` feature in the
scene graph, which would further simplify matters.