Limit to the last N records, instead of lines
Due to the fact that we cannot keep QPlainTextEdit.maximumBlockCount
active while inserting but rather have to trigger it after each edit,
this does not make a performance difference for the document clipping
operation.

Furthermore, it simplifies some of the operations (such as clipping
selections) and will simplify the optimized log rebuilding operation.