Turn shell from dockwidget into a normal dialog
The behaviour to fit the shell into the bottom of the mainwindow was
always kind-of weird and usually resulted in having to resize the
mainwindow manually afterwards. This is now made easier.

Furthermore, it was the only dockwidget left and there is not really
anything about this widget to warrant its special role. Making it a
standard dialog always frees the way to open the widget without an
active mainwindow.